move example to other file
